fix: tb_meritz_sf_damage_area_code

+++ storage/db-core/src/main/java/com/nexsol/tpa/storage/db/core/MeritzAreaCodeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MeritzAreaCodeRepository extends JpaRepository<MeritzAreaCodeEntity, Long> {
+public interface MeritzAreaCodeRepository extends JpaRepository<MeritzAreaCodeEntity, Integer> {
 
     Optional<MeritzAreaCodeEntity> findFirstByCityText1AndCityText2(String cityText1, String cityText2);
